+++ src/routes/admin/course.test.js
@@ -1,6 +1,6 @@
 import httpCodes from 'http-codes';
 import multiparty from 'multiparty';
-import { importCourse } from './Course';
+import { importCourse } from './course';
 import { handleError } from '../../lib/apiErrorHandler';
 import { csvImport } from '../../services/course';
 import { bufferFromPath } from './helpers';
